--- a/asv_bench/benchmarks/reshape.py
+++ b/asv_bench/benchmarks/reshape.py
@@ -111,7 +111,7 @@ class Unstack:
             values = np.take(list(string.ascii_letters), indices)
             values = [pd.Categorical(v) for v in values.T]
 
-        self.df = DataFrame(values, index, columns)
+        self.df = DataFrame({i: cat for i, cat in enumerate(values)}, index, columns)
         self.df2 = self.df.iloc[:-1]
 
     def time_full_product(self, dtype):