```
from asv_bench.benchmarks.reshape import *
self = ReshapeExtensionDtype()
self.setup("Period[s]")

%timeit self.time_unstack_fast("Period[s]")
23.2 ms ± 1.15 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)   # <- master
9.37 ms ± 128 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)  # <- PR
```